+++ contracts/ReClammPool.sol
@@ -17,12 +17,12 @@ import { Version } from "@balancer-labs/v3-solidity-utils/contracts/helpers/Vers
 import { PoolInfo } from "@balancer-labs/v3-pool-utils/contracts/PoolInfo.sol";
 import { BaseHooks } from "@balancer-labs/v3-vault/contracts/BaseHooks.sol";
 
-import { SqrtQ0State, AclAmmMath } from "./lib/AclAmmMath.sol";
-import { AclAmmPoolParams, IAclAmmPool } from "./interfaces/IAclAmmPool.sol";
+import { SqrtQ0State, ReClammMath } from "./lib/ReClammMath.sol";
+import { ReClammPoolParams, IReClammPool } from "./interfaces/IReClammPool.sol";
 
-contract AclAmmPool is
+contract ReClammPool is
     IUnbalancedLiquidityInvariantRatioBounds,
-    IAclAmmPool,
+    IReClammPool,
     BalancerPoolToken,
     PoolInfo,
     BasePoolAuthentication,
@@ -47,7 +47,7 @@ contract AclAmmPool is
     uint256[] private _virtualBalances;
 
     constructor(
-        AclAmmPoolParams memory params,
+        ReClammPoolParams memory params,
         IVault vault
     )
         BalancerPoolToken(vault, params.name, params.symbol)
@@ -60,13 +60,13 @@ contract AclAmmPool is
         _sqrtQ0State.endSqrtQ0 = params.sqrtQ0;
         _setCenterednessMargin(params.centerednessMargin);
 
-        emit AclAmmPoolInitialized(params.increaseDayRate, params.sqrtQ0, params.centerednessMargin);
+        emit ReClammPoolInitialized(params.increaseDayRate, params.sqrtQ0, params.centerednessMargin);
     }
 
     /// @inheritdoc IBasePool
     function computeInvariant(uint256[] memory balancesScaled18, Rounding rounding) public view returns (uint256) {
         return
-            AclAmmMath.computeInvariant(
+            ReClammMath.computeInvariant(
                 balancesScaled18,
                 _virtualBalances,
                 _c,
@@ -87,7 +87,7 @@ contract AclAmmPool is
     /// @inheritdoc IBasePool
     function onSwap(PoolSwapParams memory request) public virtual returns (uint256) {
         // Calculate virtual balances
-        (uint256[] memory virtualBalances, bool changed) = AclAmmMath.getVirtualBalances(
+        (uint256[] memory virtualBalances, bool changed) = ReClammMath.getVirtualBalances(
             request.balancesScaled18,
             _virtualBalances,
             _c,
@@ -108,7 +108,7 @@ contract AclAmmPool is
         // Calculate swap result
         if (request.kind == SwapKind.EXACT_IN) {
             return
-                AclAmmMath.calculateOutGivenIn(
+                ReClammMath.calculateOutGivenIn(
                     request.balancesScaled18,
                     _virtualBalances,
                     request.indexIn,
@@ -117,7 +117,7 @@ contract AclAmmPool is
                 );
         } else {
             return
-                AclAmmMath.calculateInGivenOut(
+                ReClammMath.calculateInGivenOut(
                     request.balancesScaled18,
                     _virtualBalances,
                     request.indexIn,
@@ -150,7 +150,7 @@ contract AclAmmPool is
         bytes memory
     ) public override onlyVault returns (bool) {
         _lastTimestamp = block.timestamp;
-        _setVirtualBalances(AclAmmMath.initializeVirtualBalances(balancesScaled18, _calculateCurrentSqrtQ0()));
+        _setVirtualBalances(ReClammMath.initializeVirtualBalances(balancesScaled18, _calculateCurrentSqrtQ0()));
         return true;
     }
 
@@ -212,22 +212,22 @@ contract AclAmmPool is
         return _MAX_INVARIANT_RATIO;
     }
 
-    /// @inheritdoc IAclAmmPool
+    /// @inheritdoc IReClammPool
     function getLastVirtualBalances() external view returns (uint256[] memory) {
         return _getLastVirtualBalances();
     }
 
-    /// @inheritdoc IAclAmmPool
+    /// @inheritdoc IReClammPool
     function getLastTimestamp() external view returns (uint256) {
         return _lastTimestamp;
     }
 
-    /// @inheritdoc IAclAmmPool
+    /// @inheritdoc IReClammPool
     function getCurrentSqrtQ0() external view override returns (uint256) {
         return _calculateCurrentSqrtQ0();
     }
 
-    /// @inheritdoc IAclAmmPool
+    /// @inheritdoc IReClammPool
     function setSqrtQ0(
         uint256 newSqrtQ0,
         uint256 startTime,
@@ -259,7 +259,7 @@ contract AclAmmPool is
         SqrtQ0State memory sqrtQ0State = _sqrtQ0State;
 
         return
-            AclAmmMath.calculateSqrtQ0(
+            ReClammMath.calculateSqrtQ0(
                 block.timestamp,
                 sqrtQ0State.startSqrtQ0,
                 sqrtQ0State.endSqrtQ0,
@@ -269,7 +269,7 @@ contract AclAmmPool is
     }
 
     function _setIncreaseDayRate(uint256 increaseDayRate) internal {
-        _c = AclAmmMath.parseIncreaseDayRate(increaseDayRate);
+        _c = ReClammMath.parseIncreaseDayRate(increaseDayRate);
     }
 
     function _setCenterednessMargin(uint256 centerednessMargin) internal {
@@ -280,7 +280,7 @@ contract AclAmmPool is
         (, , uint256[] memory balancesScaled18, ) = _vault.getPoolTokenInfo(address(this));
 
         // Calculate virtual balances
-        (virtualBalances, ) = AclAmmMath.getVirtualBalances(
+        (virtualBalances, ) = ReClammMath.getVirtualBalances(
             balancesScaled18,
             _virtualBalances,
             _c,
