+++ contracts/interfaces/IReClammPool.sol
@@ -4,8 +4,8 @@ pragma solidity ^0.8.24;
 
 import { IBasePool } from "@balancer-labs/v3-interfaces/contracts/vault/IBasePool.sol";
 
-/// @dev Struct with data for deploying a new AclAmmPool.
-struct AclAmmPoolParams {
+/// @dev Struct with data for deploying a new ReClammPool.
+struct ReClammPoolParams {
     string name;
     string symbol;
     string version;
@@ -14,7 +14,7 @@ struct AclAmmPoolParams {
     uint256 centerednessMargin;
 }
 
-interface IAclAmmPool is IBasePool {
+interface IReClammPool is IBasePool {
     /// @dev Indicates that the start time is after the end time.
     error GradualUpdateTimeTravel(uint256 resolvedStartTime, uint256 endTime);
 
@@ -22,7 +22,7 @@ interface IAclAmmPool is IBasePool {
     error NotImplemented();
 
     event SqrtQ0Updated(uint256 startSqrtQ0, uint256 endSqrtQ0, uint256 startTime, uint256 endTime);
-    event AclAmmPoolInitialized(uint256 increaseDayRate, uint256 sqrtQ0, uint256 centernessMargin);
+    event ReClammPoolInitialized(uint256 increaseDayRate, uint256 sqrtQ0, uint256 centernessMargin);
 
     function getLastVirtualBalances() external view returns (uint256[] memory virtualBalances);
