+++ test/foundry/ReClammPoolVirtualBalances.t.sol
@@ -9,12 +9,12 @@ import { FixedPoint } from "@balancer-labs/v3-solidity-utils/contracts/math/Fixe
 import { ArrayHelpers } from "@balancer-labs/v3-solidity-utils/contracts/test/ArrayHelpers.sol";
 import { Rounding } from "@balancer-labs/v3-interfaces/contracts/vault/VaultTypes.sol";
 
-import { BaseAclAmmTest } from "./utils/BaseAclAmmTest.sol";
-import { AclAmmPool } from "../../contracts/AclAmmPool.sol";
-import { AclAmmMath } from "../../contracts/lib/AclAmmMath.sol";
-import { IAclAmmPool } from "../../contracts/interfaces/IAclAmmPool.sol";
+import { BaseReClammTest } from "./utils/BaseReClammTest.sol";
+import { ReClammPool } from "../../contracts/ReClammPool.sol";
+import { ReClammMath } from "../../contracts/lib/ReClammMath.sol";
+import { IReClammPool } from "../../contracts/interfaces/IReClammPool.sol";
 
-contract AclAmmPoolVirtualBalancesTest is BaseAclAmmTest {
+contract ReClammPoolVirtualBalancesTest is BaseReClammTest {
     using FixedPoint for uint256;
     using ArrayHelpers for *;
 
@@ -32,9 +32,9 @@ contract AclAmmPoolVirtualBalancesTest is BaseAclAmmTest {
     function testInitialParams() public view {
         uint256[] memory virtualBalances = _calculateVirtualBalances();
 
-        uint256[] memory curentVirtualBalances = AclAmmPool(pool).getLastVirtualBalances();
+        uint256[] memory curentVirtualBalances = ReClammPool(pool).getLastVirtualBalances();
 
-        assertEq(AclAmmPool(pool).getCurrentSqrtQ0(), sqrtQ0(), "Invalid sqrtQ0");
+        assertEq(ReClammPool(pool).getCurrentSqrtQ0(), sqrtQ0(), "Invalid sqrtQ0");
         assertEq(curentVirtualBalances[0], virtualBalances[0], "Invalid virtual A balance");
         assertEq(curentVirtualBalances[1], virtualBalances[1], "Invalid virtual B balance");
     }
@@ -60,11 +60,11 @@ contract AclAmmPoolVirtualBalancesTest is BaseAclAmmTest {
         setInitialBalances(newInitialBalances[0], newInitialBalances[1]);
         (address firstPool, address secondPool) = _createNewPool();
 
-        assertEq(AclAmmPool(firstPool).getCurrentSqrtQ0(), sqrtQ0(), "Invalid sqrtQ0 for firstPool");
-        assertEq(AclAmmPool(secondPool).getCurrentSqrtQ0(), sqrtQ0(), "Invalid sqrtQ0 for newPool");
+        assertEq(ReClammPool(firstPool).getCurrentSqrtQ0(), sqrtQ0(), "Invalid sqrtQ0 for firstPool");
+        assertEq(ReClammPool(secondPool).getCurrentSqrtQ0(), sqrtQ0(), "Invalid sqrtQ0 for newPool");
 
-        uint256[] memory curentFirstPoolVirtualBalances = AclAmmPool(firstPool).getLastVirtualBalances();
-        uint256[] memory curentNewPoolVirtualBalances = AclAmmPool(secondPool).getLastVirtualBalances();
+        uint256[] memory curentFirstPoolVirtualBalances = ReClammPool(firstPool).getLastVirtualBalances();
+        uint256[] memory curentNewPoolVirtualBalances = ReClammPool(secondPool).getLastVirtualBalances();
 
         if (diffCoefficient > 0) {
             assertGt(
@@ -98,8 +98,8 @@ contract AclAmmPoolVirtualBalancesTest is BaseAclAmmTest {
         setSqrtQ0(newSqrtQ0);
         (address firstPool, address secondPool) = _createNewPool();
 
-        uint256[] memory curentFirstPoolVirtualBalances = AclAmmPool(firstPool).getLastVirtualBalances();
-        uint256[] memory curentNewPoolVirtualBalances = AclAmmPool(secondPool).getLastVirtualBalances();
+        uint256[] memory curentFirstPoolVirtualBalances = ReClammPool(firstPool).getLastVirtualBalances();
+        uint256[] memory curentNewPoolVirtualBalances = ReClammPool(secondPool).getLastVirtualBalances();
 
         if (newSqrtQ0 > initialSqrtQ0) {
             assertLt(
@@ -133,15 +133,15 @@ contract AclAmmPoolVirtualBalancesTest is BaseAclAmmTest {
 
         uint256 duration = 2 hours;
 
-        uint256[] memory poolVirtualBalancesBefore = AclAmmPool(pool).getLastVirtualBalances();
+        uint256[] memory poolVirtualBalancesBefore = ReClammPool(pool).getLastVirtualBalances();
 
         uint256 currentTimestamp = block.timestamp;
 
         vm.prank(admin);
-        AclAmmPool(pool).setSqrtQ0(initialSqrtQ, currentTimestamp, currentTimestamp + duration);
+        ReClammPool(pool).setSqrtQ0(initialSqrtQ, currentTimestamp, currentTimestamp + duration);
         skip(duration);
 
-        uint256[] memory poolVirtualBalancesAfter = AclAmmPool(pool).getLastVirtualBalances();
+        uint256[] memory poolVirtualBalancesAfter = ReClammPool(pool).getLastVirtualBalances();
 
         if (newSqrtQ > initialSqrtQ) {
             assertLt(
@@ -171,7 +171,7 @@ contract AclAmmPoolVirtualBalancesTest is BaseAclAmmTest {
     function testSwapExactIn_Fuzz(uint256 exactAmountIn) public {
         exactAmountIn = bound(exactAmountIn, 1e6, _INITIAL_BALANCE_A);
 
-        uint256[] memory oldVirtualBalances = AclAmmPool(pool).getLastVirtualBalances();
+        uint256[] memory oldVirtualBalances = ReClammPool(pool).getLastVirtualBalances();
         uint256 invariantBefore = _getCurrentInvariant();
 
         vm.prank(alice);
@@ -180,7 +180,7 @@ contract AclAmmPoolVirtualBalancesTest is BaseAclAmmTest {
         uint256 invariantAfter = _getCurrentInvariant();
         assertLe(invariantBefore, invariantAfter, "Invariant should not decrease");
 
-        uint256[] memory newVirtualBalances = AclAmmPool(pool).getLastVirtualBalances();
+        uint256[] memory newVirtualBalances = ReClammPool(pool).getLastVirtualBalances();
         assertEq(newVirtualBalances[0], oldVirtualBalances[0], "Virtual A balances do not match");
         assertEq(newVirtualBalances[1], oldVirtualBalances[1], "Virtual B balances do not match");
     }
@@ -197,7 +197,7 @@ contract AclAmmPoolVirtualBalancesTest is BaseAclAmmTest {
         uint256 invariantAfter = _getCurrentInvariant();
         assertLe(invariantBefore, invariantAfter, "Invariant should not decrease");
 
-        uint256[] memory currentVirtualBalances = AclAmmPool(pool).getLastVirtualBalances();
+        uint256[] memory currentVirtualBalances = ReClammPool(pool).getLastVirtualBalances();
         assertEq(currentVirtualBalances[0], virtualBalances[0], "Virtual A balances don't equal");
         assertEq(currentVirtualBalances[1], virtualBalances[1], "Virtual B balances don't equal");
     }
@@ -245,7 +245,7 @@ contract AclAmmPoolVirtualBalancesTest is BaseAclAmmTest {
 
     function _getCurrentInvariant() internal view returns (uint256) {
         (, , uint256[] memory balances, ) = vault.getPoolTokenInfo(pool);
-        return AclAmmPool(pool).computeInvariant(balances, Rounding.ROUND_DOWN);
+        return ReClammPool(pool).computeInvariant(balances, Rounding.ROUND_DOWN);
     }
 
     function _calculateVirtualBalances() internal view returns (uint256[] memory virtualBalances) {
