+++ test/foundry/utils/ReClammPoolContractsDeployer.sol
@@ -8,42 +8,42 @@ import { IVault } from "@balancer-labs/v3-interfaces/contracts/vault/IVault.sol"
 
 import { BaseContractsDeployer } from "@balancer-labs/v3-solidity-utils/test/foundry/utils/BaseContractsDeployer.sol";
 
-import { AclAmmPoolFactory } from "../../../contracts/AclAmmPoolFactory.sol";
+import { ReClammPoolFactory } from "../../../contracts/ReClammPoolFactory.sol";
 
 /**
  * @dev This contract contains functions for deploying mocks and contracts related to the "Acl Amm Pool". These
  * functions should have support for reusing artifacts from the hardhat compilation.
  */
-contract AclAmmPoolContractsDeployer is BaseContractsDeployer {
+contract ReClammPoolContractsDeployer is BaseContractsDeployer {
     string private artifactsRootDir = "artifacts/";
 
     constructor() {
         // if this external artifact path exists, it means we are running outside of this repo
-        if (vm.exists("artifacts/@balancer-labs/v3-pool-aclamm/")) {
-            artifactsRootDir = "artifacts/@balancer-labs/v3-pool-aclamm/";
+        if (vm.exists("artifacts/@balancer-labs/v3-pool-reClamm/")) {
+            artifactsRootDir = "artifacts/@balancer-labs/v3-pool-reClamm/";
         }
     }
 
-    function deployAclAmmPoolFactory(
+    function deployReClammPoolFactory(
         IVault vault,
         uint32 pauseWindowDuration,
         string memory factoryVersion,
         string memory poolVersion
-    ) internal returns (AclAmmPoolFactory) {
+    ) internal returns (ReClammPoolFactory) {
         if (reusingArtifacts) {
             return
-                AclAmmPoolFactory(
+                ReClammPoolFactory(
                     deployCode(
-                        _computeAclAmmPath(type(AclAmmPoolFactory).name),
+                        _computeReClammPath(type(ReClammPoolFactory).name),
                         abi.encode(vault, pauseWindowDuration, factoryVersion, poolVersion)
                     )
                 );
         } else {
-            return new AclAmmPoolFactory(vault, pauseWindowDuration, factoryVersion, poolVersion);
+            return new ReClammPoolFactory(vault, pauseWindowDuration, factoryVersion, poolVersion);
         }
     }
 
-    function _computeAclAmmPath(string memory name) private view returns (string memory) {
+    function _computeReClammPath(string memory name) private view returns (string memory) {
         return string(abi.encodePacked(artifactsRootDir, "contracts/", name, ".sol/", name, ".json"));
     }
 }
